// client/src/app/_hooks/issue/issueEnum.ts
export enum IssueCategoryEnum {
  ENGINE = 1,
  TRANSMISSION = 2,
  BRAKES = 3,
  ELECTRICAL = 4,
  BODY = 5,
  TIRES = 6,
  HVAC = 7,
  OTHER = 8,
}

export enum PriorityLevelEnum {
  LOW = 1,
  MEDIUM = 2,
  HIGH = 3,
  CRITICAL = 4,
}

export enum IssueStatusEnum {
  OPEN = 1,
  IN_PROGRESS = 2,
  RESOLVED = 3,
  CLOSED = 4,
  CANCELLED = 5,
}